feat(engine): home-folder-abbreviation + prompt-add-newline count
- dir now reads home-abbreviation (p10k HOME_FOLDER_ABBREVIATION, default
  '~') instead of hardcoding the tilde.
- prompt-add-newline takes a count (p10k PROMPT_ADD_NEWLINE_COUNT), so
  sparse layouts can leave more than one blank line; '#true' still parses
  as 1 and a bare number works.
- Dropped Layout::add_newline: it was parsed, serialized and never read by
  the renderer.

+++ crates/p11k-engine/src/wizard.rs
@@ -545,7 +545,7 @@ fn ask_empty_line(out: &mut io::Stdout, cfg: &mut Config) -> io::Result<Step<()>
         "提示符间距",
         &["紧凑（无空行）", "稀疏（prompt 之间空一行）"],
         cfg,
-        |c, i| c.layout.prompt_add_newline = i == 1,
+        |c, i| c.layout.prompt_add_newline = usize::from(i == 1),
     )
 }
